Artikelrabatt auf Integer geändert. Artikelliste  kann sortiert werden.

+++ ModevonMorgenSP/src/Artikelverwaltung/Accessoires.java
@@ -13,7 +13,7 @@ public class Accessoires extends Artikel {
 	 * @param Farbe Farbe des Accessoires.
 	 * @see Artikel
 	 */
-	public Accessoires(int artnr, int bestand, String Bezeichnung, String Art, String Geschlecht, String Hersteller, String Verfügbarkeit, String Notiz, String[]  Lieferanten, double Preis, double Rabatt, String Farbe) {
+	public Accessoires(int artnr, int bestand, String Bezeichnung, String Art, String Geschlecht, String Hersteller, String Verfügbarkeit, String Notiz, String[]  Lieferanten, double Preis, int Rabatt, String Farbe) {
 		super(artnr, bestand, Bezeichnung, Art, Geschlecht, Hersteller, Verfügbarkeit, Notiz, Lieferanten, Preis, Rabatt);
 		this.Farbe=Farbe;
 		Kategorie="Accessoires";
